feat(): add app-bar-setting brick Refs: NEXT_BUILDER-1433
