+++ bricks/basic-bricks/src/app-bar-setting/AppBarSetting.spec.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { shallow } from "enzyme";
+import { AppBarSetting } from "./AppBarSetting";
+
+describe("AppBarSetting", () => {
+  it("should work", () => {
+    const wrapper = shallow(<AppBarSetting />);
+    expect(wrapper.find("div")).toBeTruthy();
+  });
+});
