+++ bricks/basic-bricks/src/app-bar-setting/AppBarSetting.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import * as originAppbarSetting from "../app-bar/AppSetting/AppSetting";
+
+export function AppBarSetting(): React.ReactElement {
+  return <originAppbarSetting.AppSetting />;
+}
