+++ bricks/basic-bricks/src/app-bar-setting/index.spec.ts
@@ -1,0 +1,20 @@
+import ReactDOM from "react-dom";
+import "./";
+
+const spyOnRender = jest
+  .spyOn(ReactDOM, "render")
+  .mockImplementation(() => null);
+const unmountComponentAtNode = jest
+  .spyOn(ReactDOM, "unmountComponentAtNode")
+  .mockImplementation(() => null);
+
+describe("basic-bricks.app-bar-setting", () => {
+  it("should create a custom element", () => {
+    const element = document.createElement("basic-bricks.app-bar-setting");
+    expect(spyOnRender).not.toBeCalled();
+    document.body.appendChild(element);
+    expect(spyOnRender).toBeCalled();
+    document.body.removeChild(element);
+    expect(unmountComponentAtNode).toBeCalled();
+  });
+});
