+++ bricks/basic-bricks/src/app-bar/AppBar/AppBar.module.css
@@ -37,15 +37,3 @@
   align-items: center;
   flex-wrap: nowrap;
 }
-
-.dropdownMenuItem {
-  color: var(--app-bar-dropdown-menu-item-color);
-  &:hover {
-    background-color: var(--app-bar-dropdown-menu-item-hover-bg);
-  }
-
-  .dropdownBtn {
-    margin: -5px -12px;
-    padding: 5px 12px;
-  }
-}
